:hammer: chore: add MIT License
This commit introduces the MIT License to the project.

The MIT License is a permissive free software license,
allowing for broad freedom in how the software can be used,
modified, and distributed, provided the original copyright
and license notice are included. This helps to clarify the
terms under which the software is made available.